Rename `NewPacket` namespace to `Packet`
