+++ Mediapipe.Net.Tests/Framework/Packet/FloatArrayPacketTest.cs
@@ -4,7 +4,7 @@
 
 using System;
 using Mediapipe.Net.Framework;
-using Mediapipe.Net.Framework.NewPacket;
+using Mediapipe.Net.Framework.Packet;
 using NUnit.Framework;
 
 namespace Mediapipe.Net.Tests.Framework.NewPacket
