+++ Mediapipe.Net.Tests/Framework/Packet/ImageFramePacketTest.cs
@@ -5,7 +5,7 @@
 using System;
 using Mediapipe.Net.Framework;
 using Mediapipe.Net.Framework.Format;
-using Mediapipe.Net.Framework.NewPacket;
+using Mediapipe.Net.Framework.Packet;
 using NUnit.Framework;
 
 namespace Mediapipe.Net.Tests.Framework.NewPacket
